use chain hooks
