+++ src/contexts/ParaSync/index.tsx
@@ -2,16 +2,16 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import React, { useRef, useState } from 'react';
-import { ApiPromise, Keyring, WsProvider } from '@polkadot/api';
-import type { AnyApi, AnyJson, Sync } from 'types';
+import type { AnyJson, Sync } from 'types';
 import { getParaMeta } from 'config/paras';
 import { useConnect } from 'contexts/Connect';
 import { useApi } from 'contexts/Api';
 import { useEffectIgnoreInitial } from 'library/Hooks/useEffectIgnoreInitial';
-import { rmCommas } from '@polkadotcloud/utils';
 import BigNumber from 'bignumber.js';
 import type { ParaSyncContextInterface } from './types';
 import { defaultParaSyncContext } from './defaults';
+import { useAssetHub } from './Hooks/useAssetHub';
+import { useInterlay } from './Hooks/useInterlay';
 
 export const ParaSyncProvider = ({
   children,
@@ -31,12 +31,9 @@ export const ParaSyncProvider = ({
   // Reference whether the app has been synced.
   const isSyncingRef = useRef<Sync>('unsynced');
 
-  // Keyring instance to use for parachain account formatting.
-  const keyring = new Keyring();
-
-  // Metadata for the parachains used in this context.
-  const paraInterlay = getParaMeta('interlay');
-  const paraAssetHub = getParaMeta('assethub');
+  // Instantiate parachain hooks.
+  const assetHub = useAssetHub();
+  const interlay = useInterlay();
 
   // We need to connect to parachains and check the user's balances.
   const syncBalances = async () => {
@@ -44,27 +41,15 @@ export const ParaSyncProvider = ({
     isSyncingRef.current = 'syncing';
 
     // Sync chain states to get supported local and foreign assets.
-    const assetHubState = await getAssetHubBalances(activeAccount);
-    const interlayState = await getInterlayBalances(activeAccount);
+    const assetHubState = await assetHub.getBalances(activeAccount);
+    const interlayState = await interlay.getBalances(activeAccount);
 
-    // Format foreign asset metadata.
-    const foreignAssetsMetadata: Record<number, AnyJson> = {};
-    interlayState?.assetRegistry?.forEach(([idRaw, metadataRaw]: AnyApi) => {
-      const id = idRaw.toHuman()[0];
-      const metadata = metadataRaw.toHuman();
-
-      if (
-        paraInterlay.supportedAssets.find(
-          ({ symbol }: AnyJson) => symbol === id
-        )
-      ) {
-        foreignAssetsMetadata[id] = metadata;
-      }
-    });
-    setParaForeignAssets({
-      ...paraForeignAssets,
-      interlay: foreignAssetsMetadata,
-    });
+    if (interlayState?.foreignAssets) {
+      setParaForeignAssets({
+        ...paraForeignAssets,
+        interlay: interlayState.foreignAssets,
+      });
+    }
 
     const newParaBalances: Record<string, AnyJson> = {};
     if (interlayState) {
@@ -85,116 +70,6 @@ export const ParaSyncProvider = ({
       ...newParaBalances,
     });
     isSyncingRef.current = 'synced';
-  };
-
-  // Handler for fetching interlay balances. Connects to the interlay parachain, fetches token
-  // balances and disconnects immediately after.
-  const getAssetHubBalances = async (account: string) => {
-    if (!account) return null;
-
-    keyring.setSS58Format(paraAssetHub.ss58);
-    const { supportedAssets, endpoints } = paraAssetHub;
-
-    // Connect to interlay via new api instance.
-    const wsProvider = new WsProvider(endpoints.rpc);
-    const api = await ApiPromise.create({ provider: wsProvider });
-
-    // Fetch needed chain state.
-    const result: AnyApi[] = await Promise.all([
-      api.query.parachainInfo.parachainId(),
-      api.query.system.account(keyring.addFromAddress(account).address),
-      ...supportedAssets
-        .filter(({ key }) => key !== 'Native')
-        .map(({ key }) =>
-          api.query.assets.account(key, keyring.addFromAddress(account).address)
-        ),
-    ]);
-    await api.disconnect();
-
-    const [paraId, accountBalance, ...assetsRaw] = result;
-    const nativeBalance = accountBalance.toHuman();
-
-    const assets: AnyJson[] = [];
-    let i = 0;
-    assetsRaw?.forEach((a: AnyJson) => {
-      const asset = a.toHuman();
-      // index 0 is native asset.
-      const supportedAssetIndex = i + 1;
-      if (!asset) {
-        i++;
-        return;
-      }
-
-      assets.push({
-        ...asset,
-        key: supportedAssets[supportedAssetIndex].key,
-        symbol: supportedAssets[supportedAssetIndex].symbol,
-        balance: rmCommas(asset.balance),
-      });
-      i++;
-    });
-
-    return {
-      paraId: paraId.toString(),
-      assets: [
-        {
-          key: 'Native',
-          symbol: 'DOT',
-          free: rmCommas(nativeBalance?.data?.free || '0'),
-          frozen: rmCommas(nativeBalance?.data?.frozen || '0'),
-          reserved: rmCommas(nativeBalance?.data?.reserved || '0'),
-        },
-        ...assets,
-      ],
-    };
-  };
-
-  // Handler for fetching interlay balances. Connects to the interlay parachain, fetches token
-  // balances and disconnects immediately after.
-  const getInterlayBalances = async (account: string) => {
-    if (!account) return null;
-
-    keyring.setSS58Format(paraInterlay.ss58);
-    const { supportedAssets, endpoints } = paraInterlay;
-
-    // Connect to interlay via new api instance.
-    const wsProvider = new WsProvider(endpoints.rpc);
-    const api = await ApiPromise.create({ provider: wsProvider });
-
-    // Fetch needed chain state.
-    const [paraIdRaw, tokensRaw, assetRegistry]: AnyApi[] = await Promise.all([
-      api.query.parachainInfo.parachainId(),
-      api.query.tokens.accounts.entries(
-        keyring.addFromAddress(account).address
-      ),
-      api.query.assetRegistry.metadata.entries(),
-    ]);
-    const paraId = paraIdRaw.toString();
-
-    // Format token balances.
-    const tokens: AnyApi[] = [];
-    tokensRaw?.forEach(([keyRaw, valueRaw]: AnyApi) => {
-      const key = Object.keys(keyRaw.toHuman()[1])[0];
-      const symbol = Object.values(keyRaw.toHuman()[1])[0];
-      if (
-        supportedAssets.find(
-          (t: AnyJson) => t.key === key && t.symbol === symbol
-        )
-      ) {
-        const value = valueRaw.toHuman();
-        tokens.push({
-          ...value,
-          free: rmCommas(value.free || '0'),
-          frozen: rmCommas(value.frozen || '0'),
-          reserved: rmCommas(value.reserved || '0'),
-          key,
-          symbol,
-        });
-      }
-    });
-
-    await api.disconnect();
-    return { paraId, tokens, assetRegistry };
   };
 
   // NOTE: could make `syncBalances` cancelable and cancel when this useEffect is triggered.
